Migrate package manager classes to builtin repo
