+++ lib/ramble/ramble/pkgmankit.py
@@ -15,10 +15,11 @@ import llnl.util.filesystem
 from llnl.util.filesystem import *
 
 import ramble.language.package_manager_language
+import ramble.repository
 from ramble.language.package_manager_language import *
 from ramble.language.shared_language import *
-from ramble.package_manager import PackageManagerBase
 from ramble.software_environments import ExternalEnvironment
+from ramble.software_info import SoftwareInfo
 from ramble.spec import Spec
 from ramble.util.command_runner import (
     CommandRunner,
@@ -31,3 +32,8 @@ from ramble.util.command_runner import (
 from ramble.util.logger import logger
 from ramble.util.logger import logger as tty
 from ramble.util.output_capture import OUTPUT_CAPTURE
+
+base_class_type = ramble.repository.ObjectTypes.base_classes
+PackageManagerBase = ramble.repository.get_obj_class(
+    "package-manager-base", object_type=base_class_type
+)
