+++ lib/ramble/ramble/software_environments.py
@@ -9,10 +9,15 @@
 from collections import defaultdict
 
 import ramble.error
+import ramble.repository
 import ramble.util.colors as rucolor
 from ramble.namespace import namespace
-from ramble.package_manager import PackageManagerBase
 from ramble.util.logger import logger
+
+base_class_type = ramble.repository.ObjectTypes.base_classes
+PackageManagerBase = ramble.repository.get_obj_class(
+    "package-manager-base", object_type=base_class_type
+)
 
 SUB_INDENT = 2
 
@@ -148,7 +153,7 @@ class RenderedPackage(SoftwarePackage):
         Args:
             name (str): Name of package
             pkg_info (dict): Package info containing specs for supported package managers
-            package_manager (ramble.package_manager.PackageManagerBase):
+            package_manager (ramble.base_cls.builtin.PackageManagerBase):
                 package manager tied to this package
             spec (str): Package spec (used to install / load package)
             compiler (optional str): Name of package definition to use as compiler
@@ -367,7 +372,7 @@ class SoftwareEnvironment:
 
         Args:
             name (str): Name of the environment
-            package_manager (ramble.package_manager.PackageManagerBase):
+            package_manager (ramble.base_cls.builtin.PackageManagerBase):
                 Package manager associated with the environment
         """
 
@@ -591,7 +596,7 @@ class TemplateEnvironment(SoftwareEnvironment):
         Args:
             expander (ramble.expander.Expander): Expander object to use when rendering
             all_packages (dict): All package definitions
-            package_manager (ramble.package_manager.PackageManagerBase):
+            package_manager (ramble.base_cls.builtin.PackageManagerBase):
                 Package manager the environment is rendered with
 
         Returns:
@@ -714,7 +719,7 @@ class SoftwareEnvironments:
         information about used packages and environments.
 
         Args:
-            package_manager (ramble.package_manager.PackageManagerBase):
+            package_manager (ramble.base_cls.builtin.PackageManagerBase):
                 Reference to a package manager object
             env_name (str): Name of the rendered environment to mark as used
         """
@@ -921,7 +926,7 @@ class SoftwareEnvironments:
         Args:
             env_name (str): Name of environment needed by the experiment
             expander (ramble.expander.Expander): Expander object from the experiment
-            package_manager (ramble.package_manager.PackageManagerBase):
+            package_manager (ramble.base_cls.builtin.PackageManagerBase):
                 Package manager the environment is rendered with
 
         Returns:
