+++ lib/ramble/ramble/test/package_manager_functionality/package_manager_general.py
@@ -10,7 +10,7 @@ import os
 
 import pytest
 
-import ramble.package_manager
+import ramble.software_info
 import ramble.workspace
 from ramble.main import RambleCommand, RambleCommandError
 
@@ -67,7 +67,7 @@ def test_software_info_string():
     package = "zlib"
     version = "1.3.1"
 
-    info = ramble.package_manager.SoftwareInfo(name=package, version=version)
+    info = ramble.software_info.SoftwareInfo(name=package, version=version)
     text = info.to_version_text()
     assert package in text
     assert version in text
